Consolidate results in Cells page to CellsResults component
