+++ context/app/static/js/components/cells/CellsResults/index.js
@@ -1,0 +1,3 @@
+import CellsResults from './CellsResults';
+
+export default CellsResults;
