+++ context/app/static/js/components/cells/CellsResults/style.js
